paygenx/requirements.txt
Django
djangorestframework
faiss-cpu
requests
sentence-transformers
gunicorn
